+++ pmd-scala/src/main/java/net/sourceforge/pmd/lang/scala/ast/ScalaParser.java
@@ -2,10 +2,6 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-/*
- * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
- */
-
 package net.sourceforge.pmd.lang.scala.ast;
 
 import java.io.IOException;